Much simpler input key handling:
This assumes that every time the terminal sends some multi-byte escape
code, they're all sent at once and able to be read in a single os.read
call. So far that's been true 100% of the time, but I think this might
not be true in situations where the OS or terminal temporarily freezes
and saves up input to send in a batch. Either way, this implementation
is way simpler than the generator-function way, which was cool though.

Also, my previous idea for how to map keys to key handlers was kind of
dumb and overly limited. Now we just turn the whole key command into a
single string, and give it to the user's callback function, where they
can just switch on it. We also give them a convenient mapping of names
to escape codes, which they can use in each case, in case it's easier.